+++ dist-electron/preload.js
@@ -6,6 +6,8 @@ electron_1.contextBridge.exposeInMainWorld('electron', {
     getImageCount: (options) => electron_1.ipcRenderer.invoke('db:get-image-count', options),
     getImages: (options) => electron_1.ipcRenderer.invoke('db:get-images', options),
     getImageDetails: (id) => electron_1.ipcRenderer.invoke('db:get-image-details', id),
+    updateImageDetails: (id, updates) => electron_1.ipcRenderer.invoke('db:update-image-details', { id, updates }),
+    deleteImage: (id) => electron_1.ipcRenderer.invoke('db:delete-image', id),
     getFolders: () => electron_1.ipcRenderer.invoke('db:get-folders'),
     getKeywords: () => electron_1.ipcRenderer.invoke('db:get-keywords'),
     log: (level, message, data) => electron_1.ipcRenderer.invoke('debug:log', { level, message, data, timestamp: Date.now() }),